+++ raedamdjango/templates/core/navigation.html
@@ -30,7 +30,7 @@
                     <a class="dropdown-item" href="/admin/">
                         <i class="fa fa-fw fa-toolbox"></i>&nbsp;Admin Interface
                     </a>
-                    <a class="dropdown-item text-danger" href="#">
+                    <a class="dropdown-item text-danger" href="{% url 'logout' %}">
                         <i class="fa fa-fw fa-sign-out-alt"></i>&nbsp;Logout
                     </a>
                 </div>
